+++ R/modelling.R
@@ -672,6 +672,7 @@ rangeCheck <- function(model, minAge = NULL, maxAge = NULL, minNorm = NULL, maxN
 #' }
 #'
 #' @param data data frame of norm sample with ranking, powers and interaction of L and A
+#' @param formula prespecified formula, e. g. from an existing regression model; min and max functions will be ignored
 #' @param repetitions number of repetitions for cross validation
 #' @param norms determine norm score crossfit and R2 (if set to TRUE). The option is
 #' computationally intensive and duration increases with sample size, number of
@@ -693,9 +694,16 @@ rangeCheck <- function(model, minAge = NULL, maxAge = NULL, minNorm = NULL, maxN
 #' @export
 #' @examples
 #' # plot cross validation RMSE by number of terms up to 9 with three repetitions
-#' data <- prepareData()
-#' cnorm.cv(data, 3, max = 7, norms = FALSE)
-cnorm.cv <- function(data, repetitions = 1, norms = TRUE, min = 1, max = 12, cv = "full", pCutoff = NA, width = NA, raw = NA, group = NA, age = NA) {
+#' data <- prepareData(elfe)
+#' cnorm.cv(data, min = 3, max = 7, norms = FALSE)
+#'
+#' # cross validate prespecified formula
+#' # here, we will use the formula from a model to cross validate it and to retrieve norm RMSE
+#' # own regression functions can of course be used as well
+#' # data <- prepareData(elfe)
+#' # model <- bestModel(data)
+#' # cnorm.cv(data, formula = model$terms, repetitions = 5)
+cnorm.cv <- function(data, formula = NULL, repetitions = 1, norms = TRUE, min = 1, max = 12, cv = "full", pCutoff = NA, width = NA, raw = NA, group = NA, age = NA) {
 
   if(is.na(pCutoff)){
     if(nrow(data)<10000)
@@ -748,6 +756,7 @@ cnorm.cv <- function(data, repetitions = 1, norms = TRUE, min = 1, max = 12, cv 
   }
 
   # set up regression formulas (from bestModel function)
+  if(is.null(formula)){
   if (k == 1) {
     lmX <- formula(paste(raw, "L1 + A1 + L1A1", sep = " ~ "))
   } else if (k == 2) {
@@ -762,6 +771,11 @@ cnorm.cv <- function(data, repetitions = 1, norms = TRUE, min = 1, max = 12, cv 
   } else if (k == 6) {
     lmX <-
       formula(paste(raw, "L1 + L2 + L3 + L4 + L5 + L6 + A1 + A2 + A3 + A4 + A5 + A6 + L1A1 + L1A2 + L1A3 + L1A4 + L1A5 + L1A6 + L2A1 + L2A2 + L2A3 + L2A4 + L2A5 + L2A6 + L3A1 + L3A2 + L3A3 + L3A4 + L3A5 + L3A6 + L4A1 + L4A2 + L4A3 + L4A4 + L4A5 + L4A6 + L5A1 + L5A2 + L5A3 + L5A4 + L5A5 + L5A6 + L6A1 + L6A2 + L6A3 + L6A4 + L6A5 + L6A6", sep = " ~ "))
+  }
+  }else {
+    lmX <- formula
+    min <- length(formula)
+    max <- length(formula)
   }
 
 
@@ -828,7 +842,9 @@ cnorm.cv <- function(data, repetitions = 1, norms = TRUE, min = 1, max = 12, cv 
 
     # retrieve models coefficients for each number of terms
     for (i in min:max) {
-      cat(paste0("Repetition ", a, ", cycle ", i, "\n"))
+      if(norms && is.null(formula)){
+        cat(paste0("Repetition ", a, ", cycle ", i, "\n"))
+      }
       variables <- names(coef(subsets, id = i))
       variables <- variables[2:length(variables)] # remove '(Intercept)' variable
       reg <- paste0(raw, " ~ ", paste(variables, collapse = " + ")) # build regression formula
@@ -911,8 +927,9 @@ cnorm.cv <- function(data, repetitions = 1, norms = TRUE, min = 1, max = 12, cv 
   } else {
     par(mfrow = c(1, 1))
   }
-  tab <- data.frame(RMSE.raw.train = train.errors, RMSE.raw.test = val.errors, RMSE.raw.complete = complete.errors, R2.norm.train = r2.train, R2.norm.test = r2.test, Delta.R2.test = delta, Crossfit = r2.train / r2.test, norm.rmse = norm.rmse)
+  tab <- data.frame(RMSE.raw.train = train.errors, RMSE.raw.test = val.errors, RMSE.raw.complete = complete.errors, R2.norm.train = r2.train, R2.norm.test = r2.test, Delta.R2.test = delta, Crossfit = r2.train / r2.test, RMSE.norm.test = norm.rmse)
 
+  if(is.null(formula)){
   # plot RMSE
   plot(val.errors, pch = 19, type = "b", col = "blue", main = "Raw Score RMSE", ylab = "Root MSE", xlab = "Number of terms", ylim = c(min(train.errors, na.rm = TRUE), max(val.errors, na.rm = TRUE)))
   points(complete.errors, pch = 19, type = "b", col = "black")
@@ -936,26 +953,42 @@ cnorm.cv <- function(data, repetitions = 1, norms = TRUE, min = 1, max = 12, cv 
     # plot delta r2 test
     plot(tab$Delta.R2.test, pch = 19, type = "b", col = "black", main = "Norm Score Delta R2 in Validation", ylab = "Delta R2", xlab = "Number of terms", ylim = c(min(tab$Delta.R2.test, na.rm = TRUE), max(tab$Delta.R2.test, na.rm = TRUE)))
     abline(h = 0, col = 3, lty = 2)
+  }else{
+    tab$R2.norm.train <- NULL
+    tab$R2.norm.test <- NULL
+    tab$Delta.R2.test <- NULL
+    tab$Crossfit <- NULL
+    tab$RMSE.norm.test <- NULL
   }
 
   FirstNegative <- which(tab$Delta.R2.test <= 0)[1]
   suggest <- FirstNegative - 1
 
-  FirstNegativeNorm <- which(norm.rmse.min >= 0)[1]
-  suggestNorm <- FirstNegativeNorm - 1
-
-
   cat("\n")
   cat("The simulation yielded the following optimal settings:\n")
-  cat(paste0("\nNumber of terms with best crossfit: ", which.min((1 - tab$Crossfit)^2)))
+  if (norms){
+    cat(paste0("\nNumber of terms with best crossfit: ", which.min((1 - tab$Crossfit)^2)))
+  }
   cat(paste0("\nNumber of terms with best raw validation RMSE: ", which.min(tab$RMSE.raw.test)))
-  cat(paste0("\nNumber of terms with best norm validation R2: ", which.max(r2.test), "\n"))
-  cat(paste0("The first model with a negative Delta R2 in norm score cross validation is model ", FirstNegative, "\n"))
-  #cat(paste0("The first local minimum in norm score RMSE in cross validation is model ", FirstNegativeNorm, "\n"))
-  cat(paste0("Thus, choosing a model with ", suggest, " terms might be the best choice. For this, use the parameter 'terms = ", suggest, "' in the bestModel-function.\n"))
-  cat("\nPlease investigate the plots and the summary table, as the results might vary within a narrow range.")
-  cat("\nEspacially pay attention to RMSE.raw.test, r2.test, crossfit near 1 and where delta R2 stops to progress.")
+  if (norms){
+    cat(paste0("\nNumber of terms with best norm validation R2: ", which.max(r2.test), "\n"))
+    cat(paste0("The first model with a negative Delta R2 in norm score cross validation is model ", FirstNegative, "\n"))
+    cat(paste0("Thus, choosing a model with ", suggest, " terms might be the best choice. For this, use the parameter 'terms = ", suggest, "' in the bestModel-function.\n"))
+    cat("\nPlease investigate the plots and the summary table, as the results might vary within a narrow range.")
+    cat("\nEspacially pay attention to RMSE.raw.test, r2.test, crossfit near 1 and where delta R2 stops to progress.")
+  }
+
   cat("\n")
   cat("\n")
-  return(tab)
+  return(tab[min:max, ])
+  }else{
+    cat("\n")
+    cat("\n")
+
+    cat("Repeated cross validation with prespecified formula yielded the following results:\n")
+    cat("\n")
+    tab$Delta.R2.test <- NULL
+    return(tab[complete.cases(tab), ])
+  }
+
 }